[SV] Move name sanitization from ExportVerilog to SV dialect
* Move `resolveKeywordConflict` and `sanitizeName` from ExportVerilog to
  the SV dialect. This makes it easier to perform name sanitization at
  an earlier stage already.
