+++ include/circt/Dialect/SV/SVDialect.h
@@ -15,6 +15,7 @@
 
 #include "circt/Support/LLVM.h"
 #include "mlir/IR/Dialect.h"
+#include "llvm/ADT/StringSet.h"
 
 namespace circt {
 namespace sv {
@@ -37,6 +38,20 @@ private:
   /// Register all SV types.
   void registerTypes();
 };
+
+/// Given string \p origName, generate a new name if it conflicts with any
+/// keyword or any other name in the set \p recordNames. Use the int \p
+/// nextGeneratedNameID as a counter for suffix. Update the \p recordNames with
+/// the generated name and return the StringRef.
+llvm::StringRef resolveKeywordConflict(llvm::StringRef origName,
+                                       llvm::StringSet<> &recordNames,
+                                       size_t &nextGeneratedNameID);
+
+/// Sanitize the specified name for use in SV output. Auto-uniquifies the name
+/// through \c resolveKeywordConflict if required. If the name is empty, a
+/// unique temp name is created.
+StringRef sanitizeName(llvm::StringRef name, llvm::StringSet<> &recordNames,
+                       size_t &nextGeneratedNameID);
 
 } // namespace sv
 } // namespace circt
